Share core contract filtering logic in new base class
