+++ Common/Securities/Option/OptionFilterUniverse.cs
@@ -18,8 +18,6 @@ using System;
 using System.Collections.Generic;
 using QuantConnect.Data;
 using System.Linq;
-using System.Collections;
-using Python.Runtime;
 using QuantConnect.Securities.Option;
 
 namespace QuantConnect.Securities
@@ -27,50 +25,8 @@ namespace QuantConnect.Securities
     /// <summary>
     /// Represents options symbols universe used in filtering.
     /// </summary>
-    public class OptionFilterUniverse : IDerivativeSecurityFilterUniverse
+    public class OptionFilterUniverse : ContractSecurityFilterUniverse<OptionFilterUniverse>
     {
-        /// <summary>
-        /// Defines listed option types
-        /// </summary>
-        public enum Type : int
-        {
-            /// <summary>
-            /// Listed stock options that expire 3rd Friday of the month
-            /// </summary>
-            Standard = 1,
-
-            /// <summary>
-            /// Weeklys options that expire every week
-            /// These are options listed with approximately one week to expiration
-            /// </summary>
-            Weeklys = 2
-        }
-
-        internal IEnumerable<Symbol> _allSymbols;
-
-        /// <summary>
-        /// The underlying price data
-        /// </summary>
-        public BaseData Underlying
-        {
-            get
-            {
-                // underlying value changes over time, so accessing it makes universe dynamic
-                _isDynamic = true;
-                return _underlying;
-            }
-        }
-
-        private BaseData _underlying;
-
-        /// <summary>
-        /// True if the universe is dynamic and filter needs to be reapplied
-        /// </summary>
-        public bool IsDynamic => _isDynamic;
-
-        internal bool _isDynamic;
-
-        private Type _type = Type.Standard;
         // Fields used in relative strikes filter
         private List<decimal> _uniqueStrikes;
         private bool _refreshUniqueStrikes;
@@ -106,100 +62,12 @@ namespace QuantConnect.Securities
         }
 
         /// <summary>
-        /// Includes universe of weeklys options (if any) into selection
+        /// Determine if the given Option contract symbol is standard
         /// </summary>
         /// <returns></returns>
-        public OptionFilterUniverse IncludeWeeklys()
+        protected override bool IsStandard(Symbol symbol)
         {
-            _type |= Type.Weeklys;
-            return this;
-        }
-
-        /// <summary>
-        /// Sets universe of weeklys options (if any) as a selection
-        /// </summary>
-        /// <returns></returns>
-        public OptionFilterUniverse WeeklysOnly()
-        {
-            _type = Type.Weeklys;
-            return this;
-        }
-
-        /// <summary>
-        /// Returns universe, filtered by option type
-        /// </summary>
-        /// <returns></returns>
-        internal OptionFilterUniverse ApplyOptionTypesFilter()
-        {
-            // memoization map for ApplyOptionTypesFilter()
-            var memoizedMap = new Dictionary<DateTime, bool>();
-
-            Func<Symbol, bool> memoizedIsStandardType = symbol =>
-            {
-                var dt = symbol.ID.Date;
-
-                bool result;
-                if (memoizedMap.TryGetValue(dt, out result))
-                    return result;
-                var res = OptionSymbol.IsStandard(symbol);
-                memoizedMap[dt] = res;
-
-                return res;
-            };
-
-            _allSymbols = _allSymbols.Where(x =>
-            {
-                switch (_type)
-                {
-                    case Type.Weeklys:
-                        return !memoizedIsStandardType(x);
-                    case Type.Standard:
-                        return memoizedIsStandardType(x);
-                    case Type.Standard | Type.Weeklys:
-                        return true;
-                    default:
-                        return false;
-                }
-            }).ToList();
-
-            return this;
-        }
-
-        /// <summary>
-        /// Returns front month contract
-        /// </summary>
-        /// <returns></returns>
-        public OptionFilterUniverse FrontMonth()
-        {
-            var ordered = this.OrderBy(x => x.ID.Date).ToList();
-            if (ordered.Count == 0) return this;
-            var frontMonth = ordered.TakeWhile(x => ordered[0].ID.Date == x.ID.Date);
-
-            _allSymbols = frontMonth.ToList();
-            return this;
-        }
-
-        /// <summary>
-        /// Returns a list of back month contracts
-        /// </summary>
-        /// <returns></returns>
-        public OptionFilterUniverse BackMonths()
-        {
-            var ordered = this.OrderBy(x => x.ID.Date).ToList();
-            if (ordered.Count == 0) return this;
-            var backMonths = ordered.SkipWhile(x => ordered[0].ID.Date == x.ID.Date);
-
-            _allSymbols = backMonths.ToList();
-            return this;
-        }
-
-        /// <summary>
-        /// Returns first of back month contracts
-        /// </summary>
-        /// <returns></returns>
-        public OptionFilterUniverse BackMonth()
-        {
-            return BackMonths().FrontMonth();
+            return OptionSymbol.IsStandard(symbol);
         }
 
         /// <summary>
@@ -303,79 +171,6 @@ namespace QuantConnect.Securities
             return this;
         }
 
-        /// <summary>
-        /// Applies filter selecting options contracts based on a range of expiration dates relative to the current day
-        /// </summary>
-        /// <param name="minExpiry">The minimum time until expiry to include, for example, TimeSpan.FromDays(10)
-        /// would exclude contracts expiring in more than 10 days</param>
-        /// <param name="maxExpiry">The maxmium time until expiry to include, for example, TimeSpan.FromDays(10)
-        /// would exclude contracts expiring in less than 10 days</param>
-        /// <returns></returns>
-        public OptionFilterUniverse Expiration(TimeSpan minExpiry, TimeSpan maxExpiry)
-        {
-            if (_underlying == null)
-            {
-                return this;
-            }
-
-            if (maxExpiry > Time.MaxTimeSpan) maxExpiry = Time.MaxTimeSpan;
-
-            var minExpiryToDate = _underlying.Time.Date + minExpiry;
-            var maxExpiryToDate = _underlying.Time.Date + maxExpiry;
-
-            _allSymbols = _allSymbols
-                .Where(symbol => symbol.ID.Date >= minExpiryToDate && symbol.ID.Date <= maxExpiryToDate)
-                .ToList();
-
-            return this;
-        }
-
-        /// <summary>
-        /// Applies filter selecting options contracts based on a range of expiration dates relative to the current day
-        /// </summary>
-        /// <param name="minExpiryDays">The minimum time, expressed in days, until expiry to include, for example, 10
-        /// would exclude contracts expiring in more than 10 days</param>
-        /// <param name="maxExpiryDays">The maximum time, expressed in days, until expiry to include, for example, 10
-        /// would exclude contracts expiring in less than 10 days</param>
-        /// <returns></returns>
-        public OptionFilterUniverse Expiration(int minExpiryDays, int maxExpiryDays)
-        {
-            return Expiration(TimeSpan.FromDays(minExpiryDays), TimeSpan.FromDays(maxExpiryDays));
-        }
-
-        /// <summary>
-        /// Explicitly sets the selected contract symbols for this universe.
-        /// This overrides and and all other methods of selecting symbols assuming it is called last.
-        /// </summary>
-        /// <param name="contracts">The option contract symbol objects to select</param>
-        public OptionFilterUniverse Contracts(PyObject contracts)
-        {
-            _allSymbols = contracts.ConvertToSymbolEnumerable();
-            return this;
-        }
-
-        /// <summary>
-        /// Explicitly sets the selected contract symbols for this universe.
-        /// This overrides and and all other methods of selecting symbols assuming it is called last.
-        /// </summary>
-        /// <param name="contracts">The option contract symbol objects to select</param>
-        public OptionFilterUniverse Contracts(IEnumerable<Symbol> contracts)
-        {
-            _allSymbols = contracts.ToList();
-            return this;
-        }
-
-        /// <summary>
-        /// Sets a function used to filter the set of available contract filters. The input to the 'contractSelector'
-        /// function will be the already filtered list if any other filters have already been applied.
-        /// </summary>
-        /// <param name="contractSelector">The option contract symbol objects to select</param>
-        public OptionFilterUniverse Contracts(Func<IEnumerable<Symbol>, IEnumerable<Symbol>> contractSelector)
-        {
-            // force materialization using ToList
-            _allSymbols = contractSelector(_allSymbols).ToList();
-            return this;
-        }
 
         /// <summary>
         /// Sets universe of call options (if any) as a selection
@@ -391,31 +186,6 @@ namespace QuantConnect.Securities
         public OptionFilterUniverse PutsOnly()
         {
             return Contracts(contracts => contracts.Where(x => x.ID.OptionRight == OptionRight.Put));
-        }
-
-        /// <summary>
-        /// Instructs the engine to only filter options contracts on the first time step of each market day.
-        /// </summary>
-        public OptionFilterUniverse OnlyApplyFilterAtMarketOpen()
-        {
-            _isDynamic = false;
-            return this;
-        }
-
-        /// <summary>
-        /// IEnumerable interface method implementation
-        /// </summary>
-        public IEnumerator<Symbol> GetEnumerator()
-        {
-            return _allSymbols.GetEnumerator();
-        }
-
-        /// <summary>
-        /// IEnumerable interface method implementation
-        /// </summary>
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return _allSymbols.GetEnumerator();
         }
     }
